Day 9b, attempt 1
Also first try. It would be slightly more difficult if I tried to add the numbers in fron of the array, but there is no need to actually do that. (In particular, note that the last array consists of all zeroes by definition, so it does not matter which element you use.)