Day 10b, attempt 1
First try. This one is a bit more tricky. My approach for tracking which tiles are inside is based on the winding number of the tile. You do need to take care that the starting tile is set to the appropriate symbol, which is not easy and requires some backtracking. (I guess you could have also done two raycasts, one from the right and one from the left. Might have been easier in hindsight.) The trick with the winding number is to do the raycast slightly below the middle, i.e I do a raycast to the left at height, say, 0.25 of the tile (the precise height does not matter, only that it is strictly between 0 and 0.5). This allows you to handle the tiles with horizontal edges appropriately.