Day 13, attempt 3
This one did not go so well. The idea is to take iterate over all possible smudges and adjust the hashes appropriately. To do this easily, I adjusted the hash function so that you simple xor the correct value into the correct hash to adjust the cell. But there was a bug lurking in my logic: for each smudge I looked at the first reflection line and checked that it was not the original reflection line; however, instead I should take the first reflection line that is not the original.