Day 18b, attempt 1
This day did not go well. The final algorithm used a sweep line approach, going top to bottom. The is the additional insight that it suffices to cound the area that is inside the fine line (given by moving between the midpoints of the squares), and add the length of the outline + 1 to the area. The sweep line algorithm then is a simple event-based approach.

Also, apparently day 16 was missing before, so here it is.